requirements.txt
fastapi==0.115.2
uvicorn==0.23.2
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.6.4
pydantic_core==2.16.3
typing_extensions==4.10.0

bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.0.post1

cryptography==42.0.5
cffi==1.16.0
pycparser==2.21

anyio==4.3.0
starlette==0.37.2
greenlet==3.0.3
h11==0.14.0
click==8.1.7
